modified usable king safety
